+++ src/app/main/declarations/dsp-app-config.ts
@@ -1,5 +1,6 @@
 export class DspAppConfig {
     constructor(
-        public geonameToken: string
+        public geonameToken: string,
+        public iriBase: string
     ) { }
 }