+++ src/app/workspace/resource/properties/properties.component.ts
@@ -14,6 +14,7 @@ import {
     PermissionUtil,
     ProjectResponse,
     PropertyDefinition,
+    ReadColorValue,
     ReadLinkValue,
     ReadProject,
     ReadResource,
@@ -32,8 +33,9 @@ import { ConfirmationWithComment, DialogComponent } from 'src/app/main/dialog/di
 import { ErrorHandlerService } from 'src/app/main/error/error-handler.service';
 import { NotificationService } from 'src/app/main/services/notification.service';
 import { DspResource } from '../dsp-resource';
-import { IncomingService } from '../incoming.service';
 import { RepresentationConstants } from '../representation/file-representation';
+import { IncomingService } from '../services/incoming.service';
+import { ResourceService } from '../services/resource.service';
 import { UserService } from '../services/user.service';
 import {
     AddedEventValue,
@@ -86,6 +88,8 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
      */
     @Input() isAnnotation = false;
 
+    @Input() valueUuidToHighlight: string;
+
     /**
      * output `referredProjectClicked` of resource view component:
      * can be used to go to project page
@@ -109,6 +113,8 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
      * can be used for preview when hovering on resource
      */
     @Output() referredResourceHovered: EventEmitter<ReadLinkValue> = new EventEmitter<ReadLinkValue>();
+
+    @Output() regionColorChanged: EventEmitter<ReadColorValue> = new EventEmitter<ReadColorValue>();
 
     lastModificationDate: string;
 
@@ -140,11 +146,12 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
         @Inject(DspApiConnectionToken) private _dspApiConnection: KnoraApiConnection,
         private _dialog: MatDialog,
         private _errorHandler: ErrorHandlerService,
+        private _incomingService: IncomingService,
         private _notification: NotificationService,
+        private _resourceService: ResourceService,
         private _userService: UserService,
         private _valueOperationEventService: ValueOperationEventService,
         private _valueService: ValueService,
-        private _incomingService: IncomingService
     ) { }
 
     ngOnInit(): void {
@@ -260,8 +267,9 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
      * @param linkValue
      */
     openResource(linkValue: ReadLinkValue | string) {
-        const id = ((typeof linkValue == 'string') ? linkValue : linkValue.linkedResourceIri);
-        window.open('/resource/' + encodeURIComponent(id), '_blank');
+        const iri = ((typeof linkValue == 'string') ? linkValue : linkValue.linkedResourceIri);
+        const path = this._resourceService.getResourcePath(iri);
+        window.open('/resource' + path, '_blank');
     }
 
     previewResource(linkValue: ReadLinkValue) {
@@ -439,6 +447,9 @@ export class PropertiesComponent implements OnInit, OnChanges, OnDestroy {
                 });
             if (updatedValue instanceof ReadTextValueAsXml) {
                 this._updateStandoffLinkValue();
+            }
+            if (updatedValue instanceof ReadColorValue) {
+                this.regionColorChanged.emit();
             }
         } else {
             console.error('No properties exist for this resource');